+++ examples/openid_orcid.py
@@ -18,6 +18,8 @@ def openid_user_logged_in_function(openid_event: OpenIdUserLoggedInEvent) -> Use
         image_url=None,
         affiliation=None,
         email=email,
+        external_id=openid_event.id_token.sub,
+        external_label=email,
         user_group_uuids=[],
     )