+++ src/fair_wizard/automation/openid/handlers.py
@@ -10,10 +10,12 @@ def handle_openid_user_logged_in(openid_event: OpenIdUserLoggedInEvent) -> UserL
     :return: resulting :py:type:`UserLoginResponse` response
     """
     return AuthorizedUserResponse(
-        first_name=openid_event.id_token.other_claims['given_name'],
-        last_name=openid_event.id_token.other_claims['family_name'],
+        first_name=openid_event.id_token.other_claims.get('given_name'),
+        last_name=openid_event.id_token.other_claims.get('family_name'),
         image_url=None,
         affiliation=None,
-        email=openid_event.id_token.other_claims['email'],
+        email=openid_event.id_token.other_claims.get('email'),
+        external_id=openid_event.id_token.sub,
+        external_label=openid_event.id_token.other_claims.get('email'),
         user_group_uuids=[],
     )